+++ packages/moonstone/ToggleItem/ToggleItem.js
@@ -89,7 +89,7 @@ const ToggleItemBase = kind({
 		 *
 		 * @type {Function}
 		 */
-		onClick: PropTypes.func,
+		onTap: PropTypes.func,
 
 		/**
 		 * The handler to run when the toggle item is toggled.
@@ -135,10 +135,10 @@ const ToggleItemBase = kind({
 	},
 
 	handlers: {
-		onToggle: (ev, {onToggle, onClick, selected, disabled, value}) => {
-			if (!disabled && (onToggle || onClick)) {
+		onToggle: (ev, {onToggle, onTap, selected, disabled, value}) => {
+			if (!disabled && (onToggle || onTap)) {
 				if (onToggle) onToggle({selected: !selected, value});
-				if (onClick) onClick(ev);
+				if (onTap) onTap(ev);
 			}
 		}
 	},
@@ -175,7 +175,7 @@ const ToggleItemBase = kind({
 				role="checkbox"
 				{...rest}
 				aria-checked={selected}
-				onClick={onToggle}
+				onTap={onToggle}
 			>
 				{iconBefore}
 				{children}